requirements.txt
streamlit==1.21.0
openai==0.27.0
chromadb==0.4.6
langchain==0.0.210
sentence-transformers==2.2.2
python-dotenv==1.0.0
